fix: diff pull pocket crash (#199)
These changes avoid updating state directly in the rendering of
`PackageList`, which prevents the infinite loop.

I added a `delete` diff to the mocked data, since the issue only occurs
for `delete` or `update` diffs. I was also able to verify that this
causes the tests to fail on `dev`, but they pass on this branch.