build: release action should update ppa-build for the deb version (#322)
The release action currently updates the package version. However, the
build action does not run when this action pushes to main. This means if
you try to build the deb from ppa-build, it will be 1 version behind.

update the build action to run off of the release action (instead of on
push to main)