feat: implement account creation and invitation page (#390)
This PR introduces 2 features
- Account creation
- Invitation acceptance / rejection

--- 
Before there were 2 user states handled: Logged in / Not logged in. 
Now there are 3. Not logged in, authenticated but with no organization,
logged in (has organization).

## Account creation:
For SaaS, the user tries to log in via any method (tested and mocked
with Ubuntu One). If they have no organization associated to their
Landscape account, then they are redirected to `/create-account` page.
Where they can create a new organization for themselves, and they would
then be redirected to the dashboard. *Known issue mentioned above
applies here

For Self hosted, when the user opens Landscape for the first time, they
are redirected to the `/create-account` page. We know this by consuming
an endpoint telling us if this user has an account or not. After the
user creates their account, they are redirected to the login page and
can then enter the dashboard normally. *can consider immediately logging
the user in the dashboard when creating account

## Invitation:
The user opens the email link which redirects them to
`/accept-invitation/<id>`. E.g can be `/accept-invitation/1` If the user
is unauthenticated, we prompt them to log in (again tested and mocked
with Ubuntu One). Once they have authenticated themselves, they are
presented with the option to accept or reject an invitation to an
organization. Once they accept they are redirected inside the dashboard.
If they reject the flow ends there.



## To test:
For account creation:
1. Login page
2. Click on Sign in with Ubuntu One
3. Fill in organization form
4. End

Self-hosted
1. Login page -> redirects to /create-account (is done automatically
with the mocks and handlers of msw)
2. Fill out form
3. Navigated to Login page

For invitation
1. Manually go to `/new_dashboard/accept-invitation/1`
2. Sign in with Ubuntu One
3. Accept invitation -> Dashboard // Reject invitation -> confirmation
that rejection was done


## Untested flows:
Creating an organization with OIDC
Accepting an invite when signing in with OIDC