refactor: make account creation flow changes, make pro enablement fixes (#399)
Main changes:
## Pro enablement:
When the "attach" action the user is trying to perform has no effect on
any of the instances, we show a modal informing the user of that
information and only allow them to close the modal.

## Invitation flow / Account creation
### In SaaS:
The user only sees the "create organization" page if the url is
`landscape.canonical.com`. Otherwise they are met with a `No access`
page informing them they need an invitation to get in the organization.
### In self-hosted:
The user only sees the "create account" page if there is no existing
account coming from `GET /standalone-account`. Otherwise they are again
met with a `No access` page like in SaaS.

Invitation flow remains unchanged for both.

---

To make testing easier, you can toggle responses in the handlers
`auth.ts` and `standaloneAccounts.ts`. They are labeled for invites /
account creation and even if you want to test normally logging in with
OIDC / Ubuntu One.

I recommend mocking every msw endpoint with `/`

To test "create account" in SaaS, you would need to modify our hosts
file in your computer, and then point vite server to that url. E.g
adding 127.0.0.1 to etc/hosts and pointing it to
`landscape.canonical.com` should prompt the user to create account if
they are not associated with an organization. While any other url should
not cause that.